Move to git_utils
